fix: env_loader flaky test + coverage baseline 88.14% + minor test fixes
- fix: test_get_env_triggers_file_load stale env guard
- docs: coverage baseline updated to 88.14% (8,392 stmts, 946 missing)
- fix: iptorrents alt env vars flaky test ordering
- fix: btsow B-substring dict ordering
- misc: test assertion updates for bitru/xfsub/yihua

+++ tests/unit/test_plugin_bitru.py
@@ -154,7 +154,7 @@ class TestParseSize:
         assert self.mod._parse_size("512 KB") == 512 * 1024
 
     def test_multibyte_unit_mb_returns_zero(self):
-        assert self.mod._parse_size("750 MB") == int(750 * 1024**2)
+        assert self.mod._parse_size("750 MB") == (750 * 1024**2)
 
     def test_multibyte_unit_tb_returns_zero(self):
         assert self.mod._parse_size("1.5 TB") == int(1.5 * 1024**4)
